Added a test of l6rtcm4050.py
- Added a test of l6rtcm4050.py
- Indicating RTCM MT 4073 and RTCM MT 4050 messages in test/expect directory
- Merge librtcm.py to rtcmread.py
