+++ readme.md
@@ -1,4 +1,4 @@
-# QZS L6 Tool: quasi-zenith satellite L6-band tool, ver.0.1.0
+# QZS L6 Tool: quasi-zenith satellite L6-band tool, ver.0.1.1
 
 ![QZS L6 Tool](docs/img/qzsl6tool.png)
 
@@ -11,6 +11,7 @@
 - netcatの``nc``や、[RTKLIB](https://github.com/tomojitakasu/RTKLIB)の``str2str``などと一緒に利用することを想定しています。
 - 当初、準天頂衛星みちびき（QZS: quasi-zenith satellite）がL6周波数帯にて放送する補強メッセージ（CLASやMADOCA-PPP）の内容表示を目指していましたが、Galileo HASメッセージなども表示できるようになりました。
 - 2024年8月11日バージョンから[セマンテック・バージョニング](https://packaging.python.org/en/latest/discussions/versioning/#choosing-a-versioning-scheme)を導入しました。
+- [リリースノート](release_note.md)
 
 ## 動作環境
