+++ release_note.txt
@@ -1,0 +1,11 @@
+# Release Note on QZS L6 Tool
+
+## ver.0.1.1 (2024-08-22)
+- Added a test of l6rtcm4050.py
+- Indicating RTCM MT 4073 and RTCM MT 4050 messages in test/expect directory
+- Merge librtcm.py to rtcmread.py
+
+## ver.0.1.0 (2024-08-16)
+- Semantic versioning
+- Capable of reading MADOCA-PPP Ionospheric augmentation message
+- l6rtcm4050.py: QZS L6 raw to RTCM MT 4050 conversion
